# swagger.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: WeatherDelivery
    description: API para validar y notificar el clima para entregas pendientes
paths:
  /weather/validate:
    post:
      tags:
        - WeatherDelivery
      summary: Valida el clima para la ubicación
      description: Consulta el clima del día siguiente y envía notificación si es necesario.
      operationId: validateWeather
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeatherRequest'
        required: true
      responses:
        '200':
          description: Respuesta exitosa
          content:
            application/json:
              schema:
                type: string
                example:
                  forecast_code: 1189
                  forecast_description: Niebla moderada
                  buyer_notification: true
        '400':
          description: Solicitud inválida
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /weather/logNotification/{email}:
    get:
      tags:
        - WeatherDelivery
      operationId: getLogNotification
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
components:
  schemas:
    WeatherRequest:
      required:
        - email
        - latitude
        - longitude
      type: object
      properties:
        email:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double